Fix issue with access to 'isTransposed', make public 'size'
It turned out that 'private(set)' works well, but isn't accessible for other modules when the code is exported as a framework

Fix issue with access to 'isTransposed'. There was a 'public' access modifier on 'var' property, which enables user to change the property from outside. Create a private stored property and the public wrapper which returns the value of the private property

Make public wrapper for 'size' property which will return the value of the private stored property, as it's described above, and write a documentation for the wrapper

And also some minor refactor changes