Modularize unit tests and increase code coverage
Modularize unit tests: extract testing of array init and 'LosslessStringConvertible' members: string init and 'description' property, to separate test suites, 'ArrayInitTestSuite' and 'LosslessStringConvertibleTestSuite', respectively. This should be made because testing of all other features is rely on stability of string init, which in its turn rely on stability of array init.
So it's recommended to run test suites in the following order: ArrayInitTestSuite -> LosslessStringConvertibleTestSuite -> MainTestSuite. But you can run all the tests at once, and if they all pass than everything is okay.
If at least one test case fails run the test suites in the order described above to inspect isn't there an issue with stability of the array and/or string inits.

Also add some new asserts to some test cases to increase code coverage to 99,8%.